# Bitbucket/Swagger/SwaggerV1.yaml
openapi: 3.0.1
info:
  title: Serdiuk Mykyta BitBucket
  version: '1.0'
servers:
  - url: /api/v1
paths:
  /shipments:
    post:
      parameters:
        - name: quantity
          in: query
          description: Barcode for check
          required: false
          schema:
            type: number
      tags:
        - Shipments
      summary: Add shipment to db using bloom filter
      responses:
        '201':
          description: Success adding shipment with bloom filter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shipment'
        '500':
          description: Problem on the server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Shipments
      summary: Check barcode in DB with Bloom Filter
      parameters:
        - name: sipmentId
          in: path
          required: true
          description: Barcode for check
          schema:
            type: string
      responses:
        '200':
          description: Shipment with barcode found in DB using  bloom filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '404':
          description: Shipment with barcode not found in DB using bloom filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Problem on the server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          nullable: true
    NotFoundError:
      type: object
      properies:
        entityType:
          type: string
    Shipment:
      type: object
      properties:
        id:
          type: string